[Fix] #7 - 복약 계획 조회 및 추가 UI 로직 수정
분리 작업
